+++ orbit/src/main/java/com/babylon/orbit/ConfigReceiver.kt
@@ -16,12 +16,7 @@
 
 package com.babylon.orbit
 
-interface Middleware<STATE : Any, SIDE_EFFECT : Any> {
-    val initialState: STATE
-    val orbits: List<TransformerFunction<STATE, SIDE_EFFECT>>
-    val configuration: Config
-
-    data class Config(
-        val sideEffectCachingEnabled: Boolean = true
-    )
-}
+@OrbitDsl
+class ConfigReceiver(
+    var sideEffectCachingEnabled: Boolean = true
+)
